(GH-2997) Remove OMI, Podman, and LXD tests
Due to lack of prioritization and usage, we aren't expecting the Windows
Kerberos support, Podman transport, or LXD transport to become stable
anytime in Bolt's future. If these features are permanently
experimental, we don't need to test them as part of CI. This removes
tests for Podman, LXD, and Kerberos on Windows, and moves testing
infrastructure and setup directions to a new `dev-resources` directory
so that devs can still troubleshoot any issues that users run into.

!no-release-note