(maint) Remove `Bolt::Inventory::Inventory#validate` now that it's unused
As part of #2993, Bolt now resolves plugins and validates the resolved
inventory when the inventory is needed, instead of when instantiating
the Inventory class required to bring up the Bolt application. As part
of this, inventory groups will validate themselves when loaded, which
makes `Bolt::Inventory::Inventory#validate` not useful, and should mean
it's no longer called. This removes the function and it's tests.

!no-release-note