(GH-2997) Remove long running, duplicate apply tests
Integration tests that verify installing the Puppet agent take up to a
minute to run on nix, and a few of the tests are now duplicates (such
as verifying a task plugin and the `puppet_agent` plugin, given the
puppet_agent plugin *is* the task plugin). This removes long running
tests that seem unnecessary.