(maint) Break CI tests into more jobs
Now that the limit of concurrent Github Actions jobs for the Bolt repo
is 180 instead of 20, we can break the integration tests into more
separate jobs so that they don't take 50+ minutes to run. Running the
tests concurrently with rspec would also be an option, but many of these
tests use the same container infrastructure and we want to avoid multiple
access errors.

This moves the transport integration tests into their own CI jobs, moves
the apply-related integration tests into their own job, and moves
BoltSpec and BoltServer tests to their own job. The goal of these jobs
is to have the time for all jobs to run be less than 20 minutes to get
quicker feedback on PRs.

!no-release-note