(GH-2921) Support loading from scripts/ in BoltServer
This modifies the `project_file_metadatas` BoltServer endpoint use the
behavior currently under the `future.file_paths` feature flag in Bolt by
default.

When a user sends a file request to BoltServer, in includes a Puppet
file reference. BoltServer can't know if that file reference uses the
old, 'generic' Puppet file reference syntax `<module>/<path>` or the new
`<module>/files/<path>` syntax. To support both, BoltServer will now
first look in the old location for the given path, and if the file isn't
found it will parse the provided path to determine whether it can look
in `files/` or `scripts/` for the provided subpath. If the file isn't
found in either the old or new expected location, it will return 400 as
it did previously when a file wasn't found.

This also includes tarring the scripts directory in the plugin tarballs,
as scripts can be referenced from task metadata so need to be available
as part of the plugin tarball.

This additionally fixes a bug in the Applicator where `scripts` files
were not included in the plugins tarball to apply Puppet code, meaning
that using the scripts mount in apply blocks and manifests didn't work.

!bug

* **Allow using the scripts file mount in Puppet manifests**

  This fixes a bug where `scripts/` directory files needed for the new
  Puppet scripts mount were not included in the tarball of dependencies
  for applying manifests. Users can now use the scripts mount from Puppet
  manifests.