(GH-2977) Set main plan ID correctly
This fixes a small bug where the main plan ID was being set incorrectly.
Previously, the main plan ID was set to 1, while the fiber executor
expected the main plan ID to be 0. The main plan ID is now correctly set
to 0 as expected by the fiber executor.

This also adds documentation for running plans in verbose mode to view
error results from a parallel block.

!bug

* **Correctly set main plan ID**
  ([#2977](https://github.com/puppetlabs/bolt/issues/2977))

  Bolt now sets the main plan's ID to the correct value. Previously, the
  main plan ID was set incorrectly, causing Bolt to issue warnings
  during plan runs when it should not have.