(PE-32819) Document wrapping scripts in a plan
This adds two new documentation pages: Running scripts, and Creating a
script plan. Running scripts details running scripts using Bolt on the
command line and passing arguments to scripts. Creating a script plan
describes how to turn a script into a plan that runs that script, how to
pass arguments to the plan and through to the script, and when to use a
Task instead of a script plan.

This documentation assumes that loading from the `scripts/` directory
works by default, so this shouldn't be published until we remove the
feature flag to enable loading scripts there.

!no-release-note