Disallow aliases in Yaml config files
For details, see discussion in
https://github.com/puppetlabs/bolt/pull/3021.

Summary reasoning: the development team's preference is to not deviate
in any way from Yaml#safe_load's conservative default options.

Investigating and considering on a use-case basis the applicability or
non-applicability of the reasoning behind Yaml#safe_load's default
options is deemed not of value, or at least not a UX priority at this
time.

This commit modifies Bolt to print a meaningful error message to the
user when aliases are encountered, communicating clearly that Bolt does
not support that yaml feature.

!bug

* **Bolt now communicates clearly that YAML aliases are not supported** ([#2807](https://github.com/puppetlabs/bolt/issues/2807))

  Yaml aliases, as defined in the Yaml spec, are not supported by
  Bolt in inventory.yaml files and others. This aligns with the
  default behavior of Ruby's YAML#safe_load.