(bug-fix) Return resolved config object to inventory
When resolving inventory plugins for transport configuration the
inventory would call `resolve()` on each transport object in it's config
map (which maps the transport name "ssh" to the class
Bolt::Transport::SSH object). Previously the `resolve()` function
returned `nil` since it's last method is to call `validate`, which
caused the inventory value for the config key to be nil instead of the
class itself. This amends the `resolve()` function to return the
modified transport class so that the inventory config map is correctly
updated.

!bug

* **Correctly set inventory config when resolving plugins**

  Previously, Bolt would resolve plugins then set the config in the
  Inventory to `nil`, causing errors like `undefined method 'merge' for
  nil:NilClass`. Bolt now correctly returns the resolved config to the
  inventory.