(GH-3034) Resolve puppetdb plugins only when needed
This updates the CLI class to only load the puppetdb client and resolve
plugins for the client as needed. Previously, the puppetdb client was
loaded for every command, which would cause Bolt to error when the
config for the client used plugins that were not available to the
project (such as when installing a project's modules for the first
time).

!bug

* **Resolve puppetdb client configuration plugins as needed**
  ([#3034](https://github.com/puppetlabs/bolt/issues/3034))

  Bolt now only resolves plugins for the puppetdb client when needed.
  Previously, Bolt would attempt to resolve plugins for the puppetdb
  client for every command, which would cause Bolt to error if the
  plugin module was not available (such as when installing a project's
  modules for the first time).