(bug-fix) Error if the project doesn't have a name
This was just a small thing I noticed, when using the default Bolt
project the project doesn't have a name, and so there's no way for `bolt
policy new` to succeed, but the error doesn't say as much. This raises
an error instructing the user to create a project with a name before
creating policies.

!no-release-note