(GH-2974) Include plan docstring in `plan show` output
This updates the `bolt plan show <PLAN>` and `Get-BoltPlan -Name <PLAN>`
commands to include a plan's docstring in the output. Previously, these
commands would show just a plan's summary when a `@summary` tag was
used, or the plan's docstring if the `@summary` tag was not used. This
resulted in Bolt not showing all of a plan's documentation when using
these commands (see the `canary` plan for an example).

!bug

* **Include a plan's docstring in `plan show` output**
  ([#2974](https://github.com/puppetlabs/bolt/issues/2974))

  The `bolt plan show <PLAN>` and `Get-BoltPlan -Name <PLAN>` commands
  now include a plan's docstring in the output, even when the plan
  includes a `@summary` tag. Previously, Bolt would only display a
  plan's summary or docstring, but not both.