(GH-1527) Add `puppetdb-instances` config option
This adds a new `puppetdb-instances` config option to each of the config
layers. This option is used to configure additional named PuppetDB
instances that Bolt can connect to.

When this option is used in multiple config layers, Bolt shallow merges
the config for each PDB instance, with config at the project level
having the highest precedence.