(GH-1527) Add `--puppetdb` CLI option
This adds a new `--puppetdb` CLI option that can be used to specify a
named PuppetDB instance as the default PuppetDB instance. Changing the
default PuppetDB instance allows users to connect to a named PuppetDB
instance when one is not specified in the `puppetdb_*` and `apply` plan
functions or the `puppetdb` plugin.