(GH-3063) Support running tasks in no-op mode over pcp
Tasks run over pcp can now be run in no-op mode. Previously, running
tasks in no-op mode would cause Bolt to error with a message saying
the task does not support running in noop mode as task metadata is not
available when running over the pcp transport. Now, if the task is run
in no-op mmode, the `_noop` metaparameter will be added to the task
input before it is run.

!bug

* **Support running tasks in no-op mode over pcp**
  ([#3063](https://github.com/puppetlabs/bolt/issues/3063))

  Bolt no longer errors when running tasks in no-op mode over the pcp
  transport.