(BOLT-1585) Fix bug for Psych support
The intended logic for deterining the correct invocation of the Psych method was flawed such that the "legacy" function invocation was always used. When the new Psych version shipped with ruby 3, the positional parameter is no longer supported. This commit updates the logic to correctly determine if the named argument is available in the Psych class.