(maint) Fix race condition in file cache spec
This commit updates the spec test to ensure that cache reads and writes happen safely in bolt server's file cache. Previously the tests relied on the write to the cache executing first (because that thread was started first). This is not a safe guarantee. With this commit the cache is primed so that whether the read or write thread executes first there will be data to read. Note that the primary intent of this test is to ensure the "serial_execute" is being properly utilized to synchronise reads and writes and priming the cache does not necessarily negate the value of this test (IE creating a *new* file in a thread is not of primary importance).