(BOLT-1586) Remove `ssh` role from bolt controller in acceptance tests
This updates the acceptance test Rakefile to not add the `ssh` role to
the Bolt controller. Previously, the `ssh` role was added to the Bolt
controller, which caused some tests run on the controller to connect
over SSH. This caused problems when installing the puppet-agent package
on the controller during acceptance tests, as the
`puppet_agent::install` task added logic to not install the puppet-agent
package when a gem install of the agent is detected in
`puppet_agent >= 4.10.0`. Removing the `ssh` role results in Bolt using
the puppet gem instead of the puppet-agent package.