(maint) Fix logic for detecting suitable Psych.parse invocation
In order to support the Psych module that ships with ruby >= 2.5 it is necessary to determine if
the `filename` argument should be positional or a keyword. Previosly this logic just checked if
a parameter called `filename` existed (which it does in the supported version range), the important
check is whether or not the argument should be positional or keyword. This commit updates the logic
to explicitly check if the argument is a keyword before invoking the parse method.

!no-release-note