(maint) Properly configure bundler in bolt-server build
Previously we were setting bundler config options at the user level in
the build container, which meant they didn't get copied over to the
runtime container. That caused gems not to be found at runtime. We now
set them in the local .bundle/config file so they get copied over to the
final build.