(GH-3122) Allow larger set of characters in group and alias names
This updates the regex used to validate group and alias names to be more
permissive with the characters that can be included in the name.
Previously, group and alias names were restricted to letters, digits,
underscores, and hyphens. This updates the regex to allow any character
other than whitespace or commas (as these are used as delimiters for
target and group names when retrieving targets from inventory).

!feature

* **Allow larger set of characters in group and alias names**
  ([#3122](https://github.com/puppetlabs/bolt/issues/3122))

  Group and alias names can now include any character other than
  whitespace characters and commas.