(hack) Add `bolt script show` and `Get-BoltScript` commands
This adds new `bolt script show` and `Get-BoltScript` commands.
When used without a script specified, they list all scripts available to
the project. When a script is specified, they print the script to the
console.

These commands load scripts from the `scripts/` and `files/scripts/`
directories in a project and modules on the modulepath, similar to the
`bolt script run` and `Invoke-BoltScript` commands. It does not load
scripts that are in the `files/` directory, as Bolt does not know
whether a file in that directory is a script or a file used elsewhere in
a module.

This also adds support for filtering scripts with a substring using the
`--filter` command-line option.

!feature

* **Add `bolt script show` and `Get-BoltScript` commands**

  You can now view scripts available to a project using the new `bolt
  script show` and `Get-BoltScript` commands. These commands show
  scripts in the `scripts/` and `files/scripts/` directories for
  projects and modules. Output can be filtered with a substring using
  the `--filter` and `-Filter` options.