(hack) Add allow list for filtering scripts
This adds a new `scripts` project configuration option that allows users
to create an allowlist of scripts that are available to the project.
This option supports globbing. Any scripts that are not present in the
allowlist will not appear in `bolt script show` or `Get-BoltScript`
output, though they can still be run by users directly.

!feature

* **Add `scripts` project configuration option**

  This new configuration option allows users to create an allowlist of
  scripts that are available to the project. Glob patterns are
  supported. Any scripts not in the allowlist are hidden from `bolt
  script show` and `Get-BoltScript` output, but can still be run
  directly by users.