Work around PuppetDB bug involving large array queries
PuppetDB has degenerate performance when parsing a query that contains a
long array. When querying facts for a list of certnames, we generate a
query that enumerates all the certnames to consider. This triggers the
degenerate performance when querying for a large number (on the order of
tens of thousands) of nodes. We now follow a common pattern of breaking
this single array into many nested arrays, capping the length of any
individual array at 100 elements.

We also now use an `in` query instead of an `or` query, which doesn't
have a performance impact but is cleaner to generate.

!bug

* **Improve performance when using the 'puppetdb' plugin with many nodes**

  Using the 'puppetdb' plugin with many thousands of nodes resulted in
  degenerate performance, sometimes taking minutes to resolve simple
  queries. This is now resolved and these queries should execute
  quickly.