Update orchestrator_client dependency
Before orchestrator_client-0.5.4, the verion of faraday was not
constrained.  When resolving dependencies, bundler install
orchestrator_client-0.5.3 in order to avoid the constraint of an old
version of faraday and install the latest faraday-2.7.2.

The latest version of orchestrator_client has a constraint for faraday
to use version `~> 1.4`.

Update the minimal version of orchestrator_client to fix some CI
failures.