(maint) Pin sematic-puppet for component acceptance
The bolt runners in the component acceptance tests have ruby interpreters < 2.7. There is a new sematic puppet release that requires ruby 2.7 and the gem install commmand is not smart enough to pull in dependencies based on ruby version requirements. This commit updates the test runners to pre-install a suitable sematic puppet version before installing the bolt gem.

!no-release-note