(GH-3236) Only set ruby env vars on local transport when they exist
Previously we would attempt to use any prior environment variables captured when bolt was invoked by simply looking for those existing. This works when those exist but in the case they did not exist the environment variables were set to empty strings. This makes it impossible to load ruby code (for example to do an apply on local transport). This commit updates the local transport to only set those environment variables if they are explicitly set to something other than an empty string.

!bug

* **Only set ruby env vars for local transport when non empty strings**
  ([#3236](#3236))

  Previously when using bundled-ruby=false for the local transport and
  not expliclty setting ruby environment variables internal apply tasks
  did not work due to ruby code not being able to be loaded. Now the local
  transport only preserves ruby environment variables when they are set
  to something other than an empty string.