Improve add_facts performance
When making multiple add_facts calls to add facts to an already
present and large set of target facts performance declines
as more are added due to the initialization and copying of
the large hashes each time.

This change merges the added facts in place to the existing
facts, which provides a slight performance improvement.

!feature

* **Minor add_facts optimization**

  Deep merge new facts into existing target fact hash, instead
  of creating and copying a new hash each time.