(GH-3296) Prefer cert auth to token auth for puppetdb client
Previously regardless of using certs any puppetdb token (either read from default location OR configured in settings) would be sent in x-authentication header for puppetdb requests. In the case a cert is configured, do not include this as the puppetdb endpoint will 401 in the case a valid cert but revoked token is presented.

!bug

* **Prefer cert based auth over token for puppetdb** ([#3296](#3296))

  When both a token and cert are computed for puppetdb config, only use
  cert auth. This matches behavior of other puppetdb CLI tools.