(GH-3310) Ensure plugin code for core types is available on local transport
When using the local transport defaulting to `bundled-ruby: true` it is assumed an agent is on the target. Generally when a `puppet-agent` feature is applied to a target this is assumed to be the case. This provides the core types and providers when an catalog is applied. We explicitly do not want to pluginsync all the core types and providers when we do not have to. This commit updates the apply_catalog task to take advantage of the fact that bolt packages have the core types in them. This works by sending as a task parameter the install location of bundled module content in bolt packages. After unpacking the plugins sent to the apply_catalog task, if the task finds any modules shipped with bolt, it will add those to the `LOAD_PATH`. This allows us to continue to avoid excessive pluginsync and also support the bundled-ruby config over the local transport.

!bug

* **Ensure core types are available for apply over local transport** [#GH-3310](GH-3310)

  Previously when using `bundled-ruby: true` config on the local transport
  core types were not available during catalog application. This commit
  makes the core types available by loading the bundled bolt module content
  shipped with bolt packages if it is present on the target.