Fail if BOLT_FORGE_TOKEN not passed in.
Bolt is moving from opensource to private.  One contraint is that users must have a forge token.  A simple way to enforce this is to ensure an environment variable is present containing a valid forge token.  Environment variables are recognized as a sensible and secure method for passing credentials into applications due to their simplicity, security, and flexibility.  Therefore, I decided to pass forge token authentication via the `BOLT_FORGE_TOKEN` environment variable.

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>